+++ src/services/actions/ingredientsConstructor.js
@@ -1,0 +1,2 @@
+export const DELETE_ITEM_CONSTRUCTOR = 'DELETE_ITEM_CONSTRUCTOR';
+export const ADD_ITEM_CONSTRUCTOR = 'ADD_ITEM_CONSTRUCTOR';
