+++ src/services/actions/modal.js
@@ -1,0 +1,2 @@
+export const SELECT_ITEM = 'SELECT_ITEM';
+export const CLEAR_ITEM = 'CLEAR_ITEM';
