+++ src/services/reducers/rootReducer.js
@@ -1,8 +1,12 @@
 import { combineReducers } from 'redux';
 import { ingredientsReducer } from './ingredients';
 import { orderReducer } from './order';
+import { modalReducer } from './modal';
+import { ingredientsConstructorReducer } from './ingredientsConstructor';
 
 export const rootReducer = combineReducers({
   ingredients: ingredientsReducer,
+  ingredientConstructor: ingredientsConstructorReducer,
   order: orderReducer,
+  modal: modalReducer,
 });
